requirements.txt
numpy
pandas
scikit-learn
pillow
joblib
pytest